feat: move auxility functions to helper file
the index.js is now focused more on the chart creation and updates

refs #95
